requirements.txt
Django==2.2.10
djangorestframework==3.11.0
django-debug-toolbar==1.11
Pillow==6.0.0
django-storages==1.7.1
gunicorn==19.9.0
boto3==1.9.184
django-filter==2.2.0
